pset8/seasons/test_seasons.py
# Import the convert_minutes_to_words function from the 'seasons' module for testing
from seasons import convert_minutes_to_words

def test_seasons():
    # Test cases for the convert_minutes_to_words function

    # Test case 1: 15266880 minutes should be converted to "Fifteen million, two hundred sixty-six thousand, eight hundred eighty minutes"
    assert convert_minutes_to_words(15266880) == "Fifteen million, two hundred sixty-six thousand, eight hundred eighty minutes"

    # Test case 2: 12592800 minutes should be converted to "Twelve million, five hundred ninety-two thousand, eight hundred minutes"
    assert convert_minutes_to_words(12592800) == "Twelve million, five hundred ninety-two thousand, eight hundred minutes"

    # Test case 3: 17370720 minutes should be converted to "Seventeen million, three hundred seventy thousand, seven hundred twenty minutes"
    assert convert_minutes_to_words(17370720) == "Seventeen million, three hundred seventy thousand, seven hundred twenty minutes"

    # Test case 4: Test invalid date input, expecting "Invalid Date" as the result
    assert convert_minutes_to_words(29, 1, 2983) == "Invalid Date"

"""
The code imports the convert_minutes_to_words function from the 'seasons' module.
The test_seasons function is defined to contain test cases for the imported function.
Test case 1 checks if the function correctly converts 15266880 minutes to its word representation.
Test case 2 checks if the function correctly converts 12592800 minutes to its word representation.
Test case 3 checks if the function correctly converts 17370720 minutes to its word representation.
Test case 4 checks the behavior of the function when provided with an invalid date input (29, 1, 2983), expecting the result "Invalid Date".
These test cases serve to verify the correctness and expected behavior of the convert_minutes_to_words function for different input scenarios.
"""